Launch silent installer with and without "runNppAfterSilentInstall" param
For the new feature of Notepad++: "Add option for updating Notepad++ on exit" (https://github.com/notepad-plus-plus/notepad-plus-plus/pull/16626), Notepad++ should not be relaunched, because after closing Notepad++ then updating Notepad++, updated Notepad++ relaunch is not expected.

In this commit We modify the installer parameter on the fly, according the présence of Notepad++ instance, to meet the both needs (updating on startup & on exit).

Ref: https://github.com/notepad-plus-plus/notepad-plus-plus/pull/16626/files#r2155056335

Close #91